requirements.txt
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
codecov==2.0.22
coverage==5.1
cycler==0.10.0
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
more-itertools==8.2.0
numpy==1.18.4
packaging==20.3
pika==1.1.0
Pillow==7.1.2
pluggy==0.13.1
protobuf==3.12.1
py==1.8.1
pymongo==3.10.1
pyparsing==2.4.7
pytest==5.4.2
pytest-cov==2.8.1
python-dateutil==2.8.1
requests==2.23.0
six==1.14.0
urllib3==1.25.9
wcwidth==0.1.9
Werkzeug==1.0.1
